requirements.txt
pandas==2.3.3
mysql-connector-python==9.4.0
tabulate==0.9.0
python-dotenv==1.2.1
reportlab==4.2.5